chore: Moves List component to fs/components
