+++ packages/components/src/atoms/List/List.tsx
@@ -2,7 +2,6 @@ import React, { forwardRef } from 'react'
 import type { ElementType, HTMLAttributes } from 'react'
 
 const variantToElement = {
-  description: 'dl',
   unordered: 'ul',
   ordered: 'ol',
 }
@@ -12,27 +11,34 @@ export interface ListProps<T = HTMLElement> extends HTMLAttributes<T> {
    * ID to find this component in testing tools (e.g.: cypress, testing library, and jest).
    */
   testId?: string
-  variant?: 'description' | 'ordered' | 'unordered'
   as?: ElementType
+  variant?: 'ordered' | 'unordered'
+  marker?: boolean
 }
 
 const List = forwardRef<HTMLUListElement, ListProps>(function List(
   {
-    testId = 'store-list',
-    variant = 'unordered',
+    testId = 'fs-list',
     as: MaybeComponent,
+    variant = 'unordered',
+    marker = false,
     ...otherProps
   },
   ref
 ) {
-  const dataAttributes = {
-    'data-testid': testId,
-    'data-fs-list': variant,
-  }
 
   const Component = MaybeComponent ?? variantToElement[variant] ?? 'ul'
 
-  return <Component ref={ref} {...dataAttributes} {...otherProps} />
+  return (
+    <Component
+      ref={ref}
+      data-fs-list
+      data-testid={testId}
+      data-fs-list-variant={variant}
+      data-fs-list-marker={marker}
+      {...otherProps}
+    />
+  )
 })
 
 export default List
