--- packages/ui/src/components/atoms/List/stories/List.mdx
@@ -1,33 +1,0 @@
-import { Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import List from '../List'
-
-# List
-
-## Unordered
-
-<Canvas>
-  <Story id="atoms-list--unordered-list" />
-</Canvas>
-
-## Ordered
-
-<Canvas>
-  <Story id="atoms-list--ordered-list" />
-</Canvas>
-
-## Description
-
-<Canvas>
-  <Story id="atoms-list--description-list" />
-</Canvas>
-
-## Props
-
-<ArgsTable of={List} />
-
-## CSS Selectors
-
-```css
-[data-fs-list] {}
-[data-fs-list='(description|ordered|unordered)'] {}
-```
